+++ openhands/core/cli.py
@@ -30,7 +30,7 @@ from openhands.events.observation import (
 )
 from openhands.llm.llm import LLM
 from openhands.runtime import get_runtime_cls
-from openhands.runtime.runtime import Runtime
+from openhands.runtime.base import Runtime
 from openhands.storage import get_file_store
