+++ openhands/runtime/__init__.py
@@ -1,24 +1,26 @@
 from openhands.core.logger import openhands_logger as logger
-from openhands.runtime.e2b.sandbox import E2BBox
+from openhands.runtime.impl.e2b.sandbox import E2BBox
 
 
 def get_runtime_cls(name: str):
     # Local imports to avoid circular imports
     if name == 'eventstream':
-        from openhands.runtime.client.runtime import EventStreamRuntime
+        from openhands.runtime.impl.eventstream.eventstream_runtime import (
+            EventStreamRuntime,
+        )
 
         return EventStreamRuntime
     elif name == 'e2b':
-        from openhands.runtime.e2b.runtime import E2BRuntime
+        from openhands.runtime.impl.e2b.e2b_runtime import E2BRuntime
 
         return E2BRuntime
     elif name == 'remote':
-        from openhands.runtime.remote.runtime import RemoteRuntime
+        from openhands.runtime.impl.remote.remote_runtime import RemoteRuntime
 
         return RemoteRuntime
     elif name == 'modal':
         logger.info('Using ModalRuntime')
-        from openhands.runtime.modal.runtime import ModalRuntime
+        from openhands.runtime.impl.modal.modal_runtime import ModalRuntime
 
         return ModalRuntime
     else:
