+++ openhands/runtime/impl/e2b/sandbox.py
@@ -7,7 +7,6 @@ from e2b import Sandbox as E2BSandbox
 from e2b.sandbox.exception import (
     TimeoutException,
 )
-
 from openhands.core.config import SandboxConfig
 from openhands.core.logger import openhands_logger as logger
