+++ openhands/runtime/impl/remote/remote_runtime.py
@@ -28,9 +28,9 @@ from openhands.events.observation import (
 )
 from openhands.events.serialization import event_to_dict, observation_from_dict
 from openhands.events.serialization.action import ACTION_TYPE_TO_CLASS
+from openhands.runtime.base import Runtime
 from openhands.runtime.builder.remote import RemoteRuntimeBuilder
 from openhands.runtime.plugins import PluginRequirement
-from openhands.runtime.runtime import Runtime
 from openhands.runtime.utils.request import (
     is_404_error,
     is_503_error,
@@ -212,7 +212,7 @@ class RemoteRuntime(Runtime):
             'command': (
                 f'/openhands/micromamba/bin/micromamba run -n openhands '
                 'poetry run '
-                f'python -u -m openhands.runtime.client.client {self.port} '
+                f'python -u -m openhands.runtime.action_execution_server {self.port} '
                 f'--working-dir {self.config.workspace_mount_path_in_sandbox} '
                 f'{plugin_arg}'
                 f'--username {"openhands" if self.config.run_as_openhands else "root"} '
