+++ tests/runtime/test_bash.py
@@ -1,4 +1,4 @@
-"""Bash-related tests for the EventStreamRuntime, which connects to the RuntimeClient running in the sandbox."""
+"""Bash-related tests for the EventStreamRuntime, which connects to the ActionExecutor running in the sandbox."""
 
 import os
 
@@ -13,7 +13,7 @@ from conftest import (
 from openhands.core.logger import openhands_logger as logger
 from openhands.events.action import CmdRunAction
 from openhands.events.observation import CmdOutputObservation
-from openhands.runtime.runtime import Runtime
+from openhands.runtime.base import Runtime
 
 # ============================================================================================================================
 # Bash-specific tests
